set up parahand training env

+++ src/mjlab/entity/entity.py
@@ -687,9 +687,9 @@ class Entity:
         # Use keyframe joint positions.
         key_qpos = mj_model.key("init_state").qpos
         nq_root = 7 if not self.is_fixed_base else 0
-        default_joint_pos = torch.tensor(key_qpos[nq_root:], device=device)[
-          None
-        ].repeat(nworld, 1)
+        default_joint_pos = torch.tensor(
+          key_qpos[nq_root:], device=device, dtype=torch.float
+        )[None].repeat(nworld, 1)
       else:
         default_joint_pos = torch.tensor(
           resolve_expr(self.cfg.init_state.joint_pos, self.joint_names, 0.0),
